fixed logged checking

+++ src/app/services/supabase.service.ts
@@ -7,34 +7,25 @@ import {
   PostgrestError,
   Session,
   SupabaseClient,
-  User,
 } from '@supabase/supabase-js';
 import { environment } from '@env/environment';
-import { BehaviorSubject } from 'rxjs';
-import { PokemonId } from '@customTypes/pokemon';
-export interface Profile {
-  id?: string;
-  username: string;
-  website: string;
-  avatar_url: string;
-}
+import { FavoriteData, PokemonId } from '@customTypes/pokemon';
+
+const noUserError = () =>
+  new PostgrestError({ message: 'No User!', code: '', details: '', hint: '' });
+
 @Injectable({
   providedIn: 'root',
 })
 export class SupabaseService {
   private supabase: SupabaseClient;
   _session: AuthSession | null = null;
-  isLogged: BehaviorSubject<boolean> = new BehaviorSubject(false);
 
   constructor() {
     this.supabase = createClient(
       environment.supabaseUrl,
       environment.supabaseKey
     );
-
-    this.authChanges((_, session) => {
-      this.isLogged.next(!!session);
-    });
   }
 
   get session() {
@@ -44,12 +35,14 @@ export class SupabaseService {
     return this._session;
   }
 
-  profile(user: User) {
-    return this.supabase
-      .from('profiles')
-      .select(`username, website, avatar_url`)
-      .eq('id', user.id)
-      .single();
+  async userId(): Promise<string | null> {
+    return this.supabase.auth
+      .getSession()
+      .then(({ data }) => data.session?.user.id || null);
+  }
+
+  async isLogged(): Promise<boolean> {
+    return this.supabase.auth.getSession().then(({ data }) => !!data.session);
   }
 
   async login(email: string): Promise<AuthError | null> {
@@ -69,27 +62,48 @@ export class SupabaseService {
 
   async signOut(): Promise<AuthError | null> {
     const { error } = await this.supabase.auth.signOut();
-
-    if (!error) {
-      this.isLogged.next(false);
-    }
-
     return error;
   }
 
-  async addFavoritePokemon(id: PokemonId): Promise<PostgrestError | null> {
-    const { error } = await this.supabase
-      .from('favorites')
-      .insert({ user_id: this.session?.user.id, pokemon_id: id });
+  async addFavoritePokemon(
+    pokemon_id: PokemonId
+  ): Promise<PostgrestError | null> {
+    return this.userId().then(async (userId: string | null) => {
+      if (!userId) return noUserError();
 
-    return error;
+      const { error } = await this.supabase
+        .from('favorites')
+        .insert({ user_id: userId, pokemon_id });
+
+      return error;
+    });
   }
 
-  async removeFavoritePokemon(id: PokemonId): Promise<PostgrestError | null> {
-    const { error } = await this.supabase
-      .from('favorites')
-      .delete()
-      .match({ user_id: this.session?.user.id, pokemon_id: id });
-    return error;
+  async removeFavoritePokemon(
+    pokemon_id: PokemonId
+  ): Promise<PostgrestError | null> {
+    return this.userId().then(async (userId: string | null) => {
+      if (!userId) return noUserError();
+
+      const { error } = await this.supabase
+        .from('favorites')
+        .delete()
+        .match({ user_id: userId, pokemon_id });
+
+      return error;
+    });
+  }
+
+  async getFavoritePokemons(): Promise<FavoriteData[]> {
+    return this.userId().then(async (userId: string | null) => {
+      if (!userId) return [];
+
+      const { data } = await this.supabase
+        .from('favorites')
+        .select('pokemon_id')
+        .match({ user_id: userId });
+
+      return data || [];
+    });
   }
 }
